Fix state resizing bug in reallm.cpp

+++ src/reallm.cpp
@@ -682,7 +682,7 @@ void main()
     GetProjExtState(0, "ak5k", "reallm_sz", buf, BUFSIZ);
     auto state_size = buf[0] != '\0' ? std::stoi(buf) : 0;
     static std::string state;
-    if (state_size > state.size())
+    if (state_size > (int)state.size())
         state.resize(state_size);
     GetProjExtState(0, "ak5k", "reallm", &state[0], state_size);
     fx_set_prev.clear();
